Add files via upload
The file containing the python code. This is the "original" file - the final saved and documented file before finishing 8th grade. The other versions aren't saved because I did not save them or use anything like GitHub before, which was a big mistake. They are a few actually, but I don't plan to include them.